MINOR: gui updates
