+++ peptdeep/webui/model_ui.py
@@ -2,26 +2,6 @@ import streamlit as st
 from peptdeep.settings import global_settings
 import multiprocessing
 
-def nce_search():
-    grid_nce_first = st.number_input('Start NCE for grid NCE search',value = global_settings['model_mgr']['transfer']['grid_nce_first']*1.0,step = 1.0)
-    global_settings['model_mgr']['transfer']['grid_nce_first'] = grid_nce_first
-    grid_nce_last = st.number_input('End NCE for grid NCE search',min_value = grid_nce_first, value = global_settings['model_mgr']['transfer']['grid_nce_last']*1.0,step = 1.0)
-    global_settings['model_mgr']['transfer']['grid_nce_last'] = grid_nce_last
-    grid_nce_step = st.number_input('Step NCE for grid NCE search', value = global_settings['model_mgr']['transfer']['grid_nce_step']*1.0,step = 1.0)
-    global_settings['model_mgr']['transfer']['grid_nce_step'] = grid_nce_step
-
-    grid_instrument = st.multiselect('Instruments for grid NCE search', (
-        global_settings['model_mgr']['instrument_group']
-    ),default = global_settings['model_mgr']['transfer']['grid_instrument']) 
-    global_settings['model_mgr']['transfer']['grid_instrument'] = grid_instrument
-
-def fine_tune():
-    epoch_ms2 = st.number_input('Epoch to tune MS2 model', value = global_settings['model_mgr']['transfer']['epoch_ms2'])
-    global_settings['model_mgr']['transfer']['epoch_ms2'] = epoch_ms2
-    epoch_rt_ccs = st.number_input('Epoch to tune RT and CCS models', value = global_settings['model_mgr']['transfer']['epoch_rt_ccs'])
-    global_settings['model_mgr']['transfer']['epoch_rt_ccs'] = epoch_rt_ccs
-
-
 def predict():
     batch_size_ms2 = st.number_input('Batch size to predict MS2', value = global_settings['model_mgr']['predict']['batch_size_ms2'])
     global_settings['model_mgr']['predict']['batch_size_ms2'] = batch_size_ms2
@@ -66,12 +46,3 @@ def show():
 
     st.write('### Prediction parameters')
     predict()
-
-    st.write('### Fine-tuning parameters (for DDA rescoring only)')
-    fine_tune()
-
-    st.write('### Grid NCE and instrument search for DDA rescoring')
-    grid_nce_search = st.checkbox('Enabled', global_settings['model_mgr']['transfer']['grid_nce_search'])
-    global_settings['model_mgr']['transfer']['grid_nce_search'] = grid_nce_search
-    if grid_nce_search is True:
-        nce_search()
